:recycle: #46 複数箇所で利用するのでカスタムフックとして実装
