+++ src/components/Upload/UploadModal/SuccessMessageArea.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 
 import { AppUrl, defaultAppUrl } from '../../../constants/url';
 import useClipboardMarkdown from '../../../hooks/useClipboardMarkdown';
+import { useCopySuccess } from '../../../hooks/useCopySuccess';
 import { CopiedGithubMarkdownMessage } from '../../LgtmImages/CopiedGithubMarkdownMessage';
 
 const Wrapper = styled.div`
@@ -181,20 +182,7 @@ export const SuccessMessageArea: React.FC<Props> = ({
   appUrl,
   callback,
 }) => {
-  const [copied, setCopied] = React.useState(false);
-
-  const onCopySuccess = React.useCallback(() => {
-    if (callback) {
-      callback();
-    }
-
-    const messageDisplayTime = 1000;
-
-    setCopied(true);
-    setTimeout(() => {
-      setCopied(false);
-    }, messageDisplayTime);
-  }, [callback]);
+  const { copied, onCopySuccess } = useCopySuccess(callback);
 
   const { imageContextRef } = useClipboardMarkdown({
     onCopySuccess,
